HW2-MarketplaceAPI/catalog-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlmodel>=0.0.14
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic>=2.0.0
#datamodel-code-generator>=0.25.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
email-validator>=2.0.0
#requests>=2.28.0
